changed database tables, and modified a lot of files
