+++ app.rb
@@ -3,7 +3,8 @@ require 'json'
 require 'base64'
 
 
-require_relative 'user_dir/user_info'
+require_relative 'config/environments'
+require_relative 'models/init'
  
 module EDocuments
   # Web controller for EDocuments API
