+++ models/user.rb
@@ -1,80 +1,31 @@
 # frozen_string_literal: true
 
 require 'json'
-require 'base64'
-require 'rbnacl/libsodium'
+require 'sequel'
 
 module EDocuments
-  # Holds a full secret users
-  class User
-    STORE_DIR = 'db/'
+  # Models a user
+  class User < Sequel::Model
+    one_to_many :documents
+    plugin :association_dependencies, documents: :destroy
 
-    # Create a new user by passing in hash of data
-    def initialize(new_file)
-      @id          = new_file['id'] || new_id
-      @username     = new_file['username']
-      @name        = new_file['name']
-      @lastname       = new_file['lastname']
-      @password = new_file['password']
-    
-    end
-
-    attr_reader :id, :username, :name, :lastname, :password
-
-    def content
-      decode_content(@content)
-    end
-
-    def save
-      File.open(STORE_DIR + id + '.txt', 'w') do |file|
-        file.write(to_json)
-      end
-
-      true
-    rescue StandardError
-      false
-    end
-
-    # note: this is not the preferred format for JSON objects
-    # see: http://jsonapi.org
+    # rubocop:disable MethodLength
     def to_json(options = {})
-      JSON({ type: 'document',
-             id: @id,
-             username: @username,
-             name: @name,
-             name: @name,
-             lastname: lastname,
-             password:password}, options)
+      JSON(
+        {
+          data: {
+            type: 'user',
+            attributes: {
+              id: id,
+              name: name,
+              surname: surname,
+              email: email,
+              phone: phone,
+            }
+          }
+        }, options
+      )
     end
-
-    def self.setup
-      Dir.mkdir(STORE_DIR) unless Dir.exist? STORE_DIR
-    end
-
-    def self.find(find_id)
-      document_file = File.read(STORE_DIR + find_id + '.txt')
-      Document.new JSON.parse(document_file)
-    end
-
-    def self.all
-      Dir.glob(STORE_DIR + '*.txt').map do |filename|
-        filename.match(/#{Regexp.quote(STORE_DIR)}(.*)\.txt/)[1]
-      end
-    end
-
-    private
-
-    def new_id
-      timestamp = Time.now.to_f.to_s
-      Base64.urlsafe_encode64(RbNaCl::Hash.sha256(timestamp))[0..9]
-    end
-
-    def encode_content(content)
-      Base64.strict_encode64(content)
-    end
-
-    def decode_content(encoded_content)
-      Base64.strict_decode64(encoded_content)
-    end
+    # rubocop:enable MethodLength
   end
 end
